Update configuration - blog metadata
Updated the website configuration to be Practicalli specific

* site-title
used as the main heading in the blue theme

* author
used in footer for copyright, post author can be set in each post meta-data

* description
Unsure where this is used at present, but could be used with the theme via the
selma template

* site-url
The domain address of the deployed website.  The :blog-prefix defines the path
to the blog, which has been kept to the default `/blog` path.  Only the blog
path is used when running the local server.