Conditional topic image style
Display an image illustrating the topic of the current post at the top right of
the page, if a relevant :topic tag is set in the post meta-data.

Selma is used to check if `post.topic` is set and if so an image is displayed.
The filename of the image is the topic value followed by `-logo-name.png`.  The
same approach is taken for the alternative text.

Topic images are being created in Inkscape, including a logo and the name of the
topic with a background matching the main colour of the logo.

Reduce the width of the image in the responsive design style for smaller
screens.