Design: enabled previews, 3 blocks most visually appealing
Enabled previews so the front page of the blog shows a summary of the latest 5
posts published on the site.

This uses the theme practicalli/html/preview.html which had the topic logo added
to it in the previous commit.

Showing the first three blocks (e.g. paragraphs) of each post seems to be the
most visually appealing amount for the preview.