Refactor cryogen to take config file during build
Updated the -main function of cryogen.core to take an argument of a
configuration file, enabling a switch between test and production environments.

This fix ensures that links in the test environment point to the test
environment and not production.

The development environment is run as a server, using lein ring server, and
remains unaffected.  It simply used the production configuration in
content/config.edn

Updated the deploy script to use the default production configuration.

Added a deploy to test script to use the test configuration.